auhtentification oaut2 do

+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/services/BidListService.java
@@ -27,7 +27,9 @@ public class BidListService {
 
     public BidList delete(Integer id) {
         BidList bid = bidListRepository.findById(id).get();
-        bidListRepository.delete(bid);
+        if(bid!=null) {
+            bidListRepository.delete(bid);
+        }
         return bid;
     }
